perf(particle-effect): pause render loop when scrolled off-screen
Add an IntersectionObserver that pauses the RAF loop while the content
leaves the viewport and resumes it if the effect is still active — recovers
the offscreen-pause behaviour the old `spoiled` lib had (matters for iOS
Low Power Mode). Extract one-time GL setup into glUtils.setupGl and loop the
SIM uniforms to stay under the file-size limit.
