+++ src/components/ParticleEffect/glUtils.js
@@ -35,6 +35,24 @@ export function linkProgram(gl) {
     return program
 }
 
+// One-time GL setup: creates the mask texture (unit 0, mipmapped, clamped),
+// points the sampler at it, and configures alpha blending. Returns the texture.
+export function setupGl(gl, program, loc) {
+    const texture = gl.createTexture()
+    gl.activeTexture(gl.TEXTURE0)
+    gl.bindTexture(gl.TEXTURE_2D, texture)
+    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR)
+    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR_MIPMAP_NEAREST)
+    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_S, gl.CLAMP_TO_EDGE)
+    gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_WRAP_T, gl.CLAMP_TO_EDGE)
+    gl.useProgram(program)
+    gl.uniform1i(loc.textTexture, 0)
+    gl.clearColor(0, 0, 0, 0)
+    gl.enable(gl.BLEND)
+    gl.blendFunc(gl.SRC_ALPHA, gl.ONE_MINUS_SRC_ALPHA)
+    return texture
+}
+
 // Resolves any CSS color (or the element's computed text color) to [r,g,b] 0..1.
 export function resolveColor(input, fallbackEl) {
     const value =
